+++ commands/submit.py
@@ -5,7 +5,7 @@ from enum import Enum, auto
 from bs4 import BeautifulSoup
 from helpers.cli import yes
 from commands.web import openSubmission
-from helpers.parser import problem
+from helpers.types import problem
 from helpers.webutils import HEADERS, promptToFetch
 from helpers.programSelector import (
     formatProgramFile,
