+++ helpers/parser.py
@@ -81,16 +81,3 @@ def unify_args(command, args, configArgs):
     return [*preCommand, command, *postCommand]
 
 
-def problem(problemName: str):
-    if problemName.startswith(".\\"):
-        problemName = problemName[2:]
-    if problemName.endswith("\\"):
-        problemName = problemName[:-1]
-    if len(problemName) == 1:
-        section = getConfig().get('contest')
-        if problemName in section:
-            return section[problemName]
-    return problemName
-
-def problemList(lst: list[str]):
-    return [problem(x) for x in lst]
